requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alabaster==0.7.16
alembic==1.13.1
altair==5.1.2
amazoncaptcha==0.5.11
amqp==5.3.1
annotated-types==0.7.0
anyio==4.2.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asn1crypto==1.5.1
asttokens==2.4.1
async-generator==1.10
async-lru==2.0.4
async-timeout==5.0.1
attrs==22.2.0
autobahn==24.4.2
Automat==24.8.1
autopep8==2.0.4
Babel==2.14.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.11.1
billiard==4.2.1
bleach==6.1.0
blinker==1.6.2
blis==1.0.1
build==1.2.2.post1
CacheControl==0.14.0
cachetools==5.3.2
catalogue==2.0.10
celery==5.4.0
certifi==2022.12.7
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==2.1.1
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.3
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.19.0
coloredlogs==15.0.1
comm==0.2.1
confection==0.1.5
constantly==23.10.4
contourpy==1.0.6
cryptography==42.0.7
cycler==0.11.0
cymem==2.0.8
daphne==4.0.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
distlib==0.3.6
distro==1.9.0
dj-database-url==2.1.0
dj-rest-auth==7.0.0
Django==4.2.7
django-allauth==0.58.2
django-axes==7.0.1
django-cors-headers==4.3.0
django-environ==0.11.2
django-filter==23.3
django-ratelimit==4.1.0
django-redis==5.4.0
django-rest-auth==0.9.5
django-rest-knox==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.7.0
docutils==0.20.1
durationpy==0.9
ecdsa==0.19.0
email_validator==2.2.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et-xmlfile==1.1.0
executing==2.0.1
fake-useragent==1.5.1
Faker==33.1.0
fastapi==0.109.0
fastjsonschema==2.19.1
filelock==3.9.0
firebase-admin==6.5.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.5
flatbuffers==25.2.10
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2025.2.0
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.19.2
google-api-python-client==2.144.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
graphviz==0.20.3
grpcio==1.66.1
grpcio-status==1.66.1
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.26.0
huggingface-hub==0.29.1
humanfriendly==10.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.8.0
importlib_resources==6.5.2
incremental==24.7.2
iniconfig==2.0.0
ipykernel==6.29.0
ipython==8.20.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jiter==0.8.2
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.11
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.4
kombu==5.4.2
kubernetes==32.0.1
langchain==0.3.19
langchain-core==0.3.37
langchain-text-splitters==0.3.6
langcodes==3.4.1
langsmith==0.3.10
language_data==1.2.0
Mako==1.3.9
marisa-trie==1.2.1
markdown-it-py==3.0.0
markdownify==0.11.6
MarkupSafe==2.1.2
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.4
murmurhash==1.0.10
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
notebook==7.0.7
notebook_shim==0.2.3
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.63.2
openpyxl==3.1.2
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.15
outcome==1.2.0
overrides==7.7.0
packaging==24.2
panda==0.3.1
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
passlib==1.7.4
patsy==0.5.6
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pdfservices-sdk==4.0.0
pexpect==4.9.0
pgvector==0.3.6
Pillow==10.1.0
platformdirs==2.6.2
plotly==5.18.0
pluggy==1.5.0
posthog==3.14.2
preshed==3.0.9
priority==1.3.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf==5.28.0
psutil==5.9.8
psycopg2==2.9.10
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pulsar-client==3.6.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyChatGPT==0.4.3.3
pycodestyle==2.11.0
pycparser==2.21
pydantic==2.10.6
pydantic-settings==2.1.0
pydantic_core==2.27.2
pydeck==0.8.1b0
Pygments==2.17.1
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.0.9
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
pytesseract==0.3.13
pytest==7.4.4
pytest-asyncio==0.23.4
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
redis==5.0.1
referencing==0.31.0
regex==2024.11.6
replicate==0.4.0
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.13.0
rsa==4.9
safetensors==0.5.2
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
selenium==4.8.0
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.2.0
service-identity==24.2.0
shellingham==1.5.4
six==1.16.0
sklearn==0.0.post11
slack_sdk==3.34.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
snowflake-connector-python==3.10.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==7.3.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.25
sqlparse==0.5.1
srsly==2.4.8
stack-data==0.6.3
starlette==0.35.1
statsmodels==0.14.2
streamlit==1.28.2
sympy==1.13.1
tenacity==8.2.3
terminado==0.18.0
thinc==8.3.2
threadpoolctl==3.2.0
tinycss2==1.2.1
tkcalendar==1.6.1
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.12.5
toolz==0.12.0
torch==2.6.0
torchvision==0.21.0
tornado==6.3.3
tqdm==4.67.1
traitlets==5.14.1
transformers==4.49.0
trio==0.22.0
trio-websocket==0.9.2
ttkthemes==3.2.2
Twisted==24.11.0
txaio==23.1.1
typer==0.12.5
types-python-dateutil==2.8.19.20240106
typing_extensions==4.12.2
tzdata==2023.4
tzlocal==5.2
undetected-chromedriver==3.2.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.14
useragent==0.1.1
uvicorn==0.27.0
uvloop==0.21.0
validators==0.22.0
vine==5.1.0
virtualenv==20.17.1
wasabi==1.1.3
watchfiles==1.0.4
wcwidth==0.2.13
weasel==0.4.1
webcolors==1.13
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==2.3.6
whitenoise==6.6.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.8.2
zipp==3.17.0
zope.interface==7.2
zstandard==0.23.0
